requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Limiter==4.0.0
python-dotenv==1.0.1
requests==2.31.0
phonenumbers==8.13.27
marshmallow==3.20.2
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
tenacity==8.2.3
APScheduler==3.10.4
pytest==7.4.4
requests-mock==1.11.0
loguru==0.7.2